+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- RPC Server load test and renaming
+
 ## [0.1.0] - 2021-11-03
 
 - Increase the reconnecting websocket connection timeout.